refactor(vault-usecases): consolidate experiment edit options
